DayThree/Backend/app_b86ed9a3-cfa5-4d79-a572-890b5a4e545f/requirements.txt
django>=2.0.11
psycopg2==2.7 --no-binary psycopg2
pillow==5.1.0
django-jet
djangorestframework
django-filter
git+https://github.com/mago960806/django-rest-framework-docs.git
